feat: pretty print for withdrawal submission
